Video class improvements
